requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
packaging==24.1
pillow==10.4.0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.4
sympy==1.13.1
timm==1.0.8
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
